Changes the init sequence for the LCD to reduce the ghosting problem.
This increases the time needed for the ghosting effect to show up, but
it doesn't eliminate it. Ghosting can now barely be seen after 45s static
high contrast image. It is clearly visible after 60s, and it is glaring
after 5 minutes.

The test is, leave the 'connect to you phone' modal up for a set period
of time and then echo 4 > /sys/class/lcd/test to see the ghost image.